==> notebooks/1_preprocessing_diabetes.ipynb <==
import pandas as pd

diabetes_data = pd.read_csv("../data/diabetes/diabetes.csv")
diabetes_data.head()

1. Проверим датасет на наличие пустых ячеек:

diabetes_data.isnull().sum()

Пустых признаков в датасете нет

2. Приведем бинарные признаки к числовому формату

for column in diabetes_data.columns.drop(['Age', 'Gender', 'Class']):
    diabetes_data[column] = diabetes_data[column].map({'No' : 0, 'Yes': 1})
diabetes_data['Class'] = diabetes_data['Class'].map({'Positive' : 1, 'Negative' : 0})
diabetes_data.head()

3. Преобразуем категориальный признак

diabetes_data = pd.get_dummies(data = diabetes_data, columns = ['Gender'])
diabetes_data.head()

4. Сохраним обработанный датасет

diabetes_data.to_csv("../data/diabetes/diabetes_preprocessed.csv", sep = ',')
