==> notebooks/1_preprocessing_diamonds.ipynb <==
import pandas as pd

diamonds_data = pd.read_csv("../data/diamonds/diamonds.csv", index_col = 0)
diamonds_data.head()

diamonds_data.describe()

1. Проверим датасет на наличие пустых ячеек:

diamonds_data.isnull().sum()

Пустых признаков в датасете нет

2. В датасете есть алмазы с параметрами x, y, z равными 0. Исключим их из набора.

diamonds_data = diamonds_data.drop(diamonds_data[diamonds_data["x"] == 0].index)
diamonds_data = diamonds_data.drop(diamonds_data[diamonds_data["y"] == 0].index)
diamonds_data = diamonds_data.drop(diamonds_data[diamonds_data["z"] == 0].index)
diamonds_data.describe()

3. В датасете присутствуют три категориальных признака, избавимся от них при помощи one-hot кодирования. Предварительно скопируем данные, чтобы не изменить оригинальный датасет.

diamonds_ml = diamonds_data.copy()

diamonds_ml = pd.get_dummies(diamonds_data)
diamonds_ml.head()

4. Сохраним обработанный датасет

diamonds_ml.to_csv("../data/diamonds/diamonds_preprocessed.csv", sep = ',')
