==> notebooks/3_regression.ipynb <==
import numpy as np
import pandas as pd
from sklearn.linear_model import LinearRegression, Lasso, LassoCV, Ridge, RidgeCV, ElasticNet, ElasticNetCV
from sklearn.tree import DecisionTreeRegressor
from sklearn.metrics import mean_squared_error, mean_absolute_error
from sklearn.model_selection import train_test_split, GridSearchCV

Загрузим датасет

diamonds_data = pd.read_csv("../data/diamonds/diamonds_preprocessed.csv", index_col = 0)
diamonds_data.head()

Отделим признаки от цены

y = diamonds_data["price"]
diamonds_data.drop('price', axis =1, inplace = True)
X = diamonds_data
X

Разделим датасет на обучающую и тестовую выборки

X_train, X_test, y_train, y_test = train_test_split(X, y)

# Хелпер

def print_quality(score, mse, rmse, mae): 
    print(f"Коэффициент детерминации: {score}")
    print(f'MSE: {mse}')
    print(f'RMSE: {rmse}')
    print(f'MAE: {mae}')

# Линейная регрессия 
##### (https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LinearRegression.html)

lr = LinearRegression().fit(X_train, y_train)
lr_predict = lr.predict(X_train)
print_quality(lr.score(X, y), mean_squared_error(y_train, lr_predict), mean_squared_error(y_train, lr_predict, squared=False),
             mean_absolute_error(y_train, lr_predict))

# DTR
##### (https://scikit-learn.org/stable/modules/generated/sklearn.tree.DecisionTreeRegressor.html)

params = {'max_depth': np.arange(1, 20, 1),
          'max_features': np.arange(1, 25, 1)}
dtr = GridSearchCV(DecisionTreeRegressor(random_state=0), params).fit(X_train, y_train)
dtr.best_params_

dtr_predict = dtr.predict(X_train)
print_quality(dtr.score(X, y), mean_squared_error(y_train, dtr_predict), mean_squared_error(y_train, dtr_predict, squared=False),
             mean_absolute_error(y_train, dtr_predict))

# Lasso
##### (https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.Lasso.html#sklearn.linear_model.Lasso)

Уберем варнинги, которые появляются при обучении модели

import warnings
warnings.filterwarnings('ignore')

params = {'alpha': np.arange(0.1, 3, 0.05)}
lasso = GridSearchCV(Lasso(random_state=0), params).fit(X_train, y_train)
lasso.best_params_

lasso_predict = lasso.predict(X_train)
print_quality(lasso.score(X, y), mean_squared_error(y_train, lasso_predict), 
              mean_squared_error(y_train, lasso_predict, squared=False),
             mean_absolute_error(y_train, lasso_predict))

# Ridge
##### (https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.Ridge.html#sklearn.linear_model.Ridge)

params = {'alpha': np.arange(0.1, 4, 0.05)}
ridge = GridSearchCV(Ridge(random_state=0), params).fit(X_train, y_train)
ridge.best_params_

ridge_predict = ridge.predict(X_train)
print_quality(ridge.score(X, y), mean_squared_error(y_train, ridge_predict), 
              mean_squared_error(y_train, ridge_predict, squared=False),
             mean_absolute_error(y_train, ridge_predict))

# ElasticNet
##### (https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.ElasticNet.html)

params = {'alpha': np.arange(0.1, 3, 0.05), 'l1_ratio': np.arange(0.1, 1.1, 0.1)}
elastic = GridSearchCV(ElasticNet(random_state=0), params).fit(X_train, y_train)
elastic.best_params_

elastic_predict = elastic.predict(X_train)
print_quality(elastic.score(X, y), mean_squared_error(y_train, elastic_predict), 
              mean_squared_error(y_train, elastic_predict, squared=False),
             mean_absolute_error(y_train, elastic_predict))

# Итог

DTR оказался самым эффективным регрессором.
