==> notebooks/4_ensemble_classification.ipynb <==
import warnings
import numpy as np
import pandas as pd
from sklearn.ensemble import BaggingClassifier, GradientBoostingClassifier, StackingClassifier, BaggingRegressor, \
    GradientBoostingRegressor, StackingRegressor
from sklearn.metrics import confusion_matrix, classification_report, mean_squared_error, mean_absolute_error
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.preprocessing import MinMaxScaler
from sklearn.tree import DecisionTreeClassifier, DecisionTreeRegressor
warnings.filterwarnings('ignore')

def print_metrics(estimator, y_test, y_pred):
    print(classification_report(y_test, y_pred))
    print(estimator.score(X_test, y_test))
    print(confusion_matrix(y_test, y_pred))

# Классификация

diabetes_data = pd.read_csv("../data/diabetes/diabetes_preprocessed.csv", index_col = 0)
diabetes_data.head()

y = diabetes_data["Class"]
diabetes_data.drop('Class', axis =1, inplace = True)
X = diabetes_data
X

X_train, X_test, y_train, y_test = train_test_split(X, y, stratify=y)

scaler = MinMaxScaler()
scaler.fit(X_train[['Age']])
scaler.fit(X_test[['Age']])

## DTC

%%time
params_dtc = {'max_depth': np.arange(1,20,1)}
dtc = DecisionTreeClassifier()
dtc_base = GridSearchCV(dtc, params_dtc).fit(X_train, y_train)
dtc_base.best_params_

print_metrics(dtc_base, y_test, dtc_base.predict(X_test))

## BaggingClassifier

%%time
params_ensemble = {'n_estimators': np.arange(10,101,20),
                    'max_features': np.arange(1,24,10)}
bagging = BaggingClassifier(base_estimator=DecisionTreeClassifier(max_depth=dtc_base.best_params_['max_depth']))
model = GridSearchCV(bagging, params_ensemble).fit(X_train, y_train)
model.best_params_

print_metrics(model, y_test, model.predict(X_test))

## GradientBoostingClassifier

%%time
grad = GradientBoostingClassifier()
model = GridSearchCV(grad, params_ensemble).fit(X_train, y_train)
model.best_params_

 print_metrics(model, y_test, model.predict(X_test))

## StackingClassifier

%%time
model = StackingClassifier(estimators=[('bag',bagging), ('gbc',grad)],
                           final_estimator=dtc_base).fit(X_train, y_train)

print_metrics(model, y_test, model.predict(X_test))

## Вывод
Композиции могут значительно повлиять на время обучения моделей. На представленых выше данных композиция не дала значительного увеличения качества модели
