==> notebooks/4_ensemble_regression.ipynb <==
import warnings
import numpy as np
import pandas as pd
from sklearn.ensemble import BaggingClassifier, GradientBoostingClassifier, StackingClassifier, BaggingRegressor, \
    GradientBoostingRegressor, StackingRegressor
from sklearn.metrics import confusion_matrix, classification_report, mean_squared_error, mean_absolute_error
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.preprocessing import MinMaxScaler
from sklearn.tree import DecisionTreeClassifier, DecisionTreeRegressor
warnings.filterwarnings('ignore')

 def print_metrics(estimator, y_test, y_pred):
    print(f"Коэффициент детерминации: {estimator.score(X,y)}")
    print(f'MSE: {mean_squared_error(y_test, y_pred)}')
    print(f'RMSE: {mean_squared_error(y_test, y_pred, squared=False)}')
    print(f'MAE: {mean_absolute_error(y_test, y_pred)}')

# Регрессия

diamonds_data = pd.read_csv("../data/diamonds/diamonds_preprocessed.csv", index_col = 0)
diamonds_data.head()

y = diamonds_data["price"]
diamonds_data.drop('price', axis =1, inplace = True)
X = diamonds_data
X

X_train, X_test, y_train, y_test = train_test_split(X, y)

## DTR

%%time
params_dtr = {'max_depth': np.arange(5,16,1)}
dtr = DecisionTreeRegressor().fit(X_train, y_train)
dtr_base = GridSearchCV(dtr, params_dtr).fit(X_train, y_train)
print(dtr_base.best_params_)

print_metrics(dtr_base, y_test, dtr_base.predict(X_test))

## BaggingRegressor

params_ensemble = {'n_estimators': np.arange(10,101,20),
                    'max_features': np.arange(1,24,10)}

%%time
br = BaggingRegressor(base_estimator=DecisionTreeRegressor(max_depth=dtr_base.best_params_['max_depth']))
model = GridSearchCV(br, params_ensemble).fit(X_train, y_train)
model.best_params_

print_metrics(model, y_test, model.predict(X_test))

## GradientBoostingRegressor

%%time
gbr = GradientBoostingRegressor()
model = GridSearchCV(gbr, params_ensemble).fit(X_train, y_train)
model.best_params_

print_metrics(model, y_test, model.predict(X_test))

## StackingRegressor

 %%time
model = StackingRegressor(estimators=[('br',br), ('gbr',gbr)],
                          final_estimator=dtr_base).fit(X_train, y_train)
print_metrics(model, y_test, model.predict(X_test))

## Вывод

Композиции могут значительно повлиять на время обучения моделей. Их применение может повысить качество моделей для задачи регрессии
