==> notebooks/5_nn.ipynb <==
import pandas as pd
import numpy as np
from sklearn import datasets
from sklearn.metrics import classification_report, confusion_matrix
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from modules import mlp

# Датасет "Титаник"

## Загрузка датасета

data = pd.read_csv('../data/nn_test/titanic.csv', sep = ',')

data.head()

## Подготовка датасета

data.drop(['PassengerId', 'Name', 'Ticket', 'Cabin'], axis=1, inplace=True)
data.fillna(data.mean(), inplace=True)
data = pd.get_dummies(data=data, prefix='c', columns=['Sex', 'Embarked'])
data.head()

X = data.drop('Survived', 1).values
y = data['Survived'].values
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, stratify=y)

## Классификация с помощью модели MLP

model = mlp.MLP(mlp.sigmoid, mlp.dsigmoid, layers=[X.shape[1], 100, 100, len(np.unique(y))])

model.fit(X_train, y_train)

## Посмотрим точность

print(sum(np.argmax(model.predict(X_test), axis=1) == y_test) / y_test.shape[0])

# Датасет "IRIS"

## Загрузка датасета

iris = datasets.load_iris()
iris

## Подготовка датасета

X = iris.data
y = iris.target

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, stratify=y)

scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

## Классификация с помощью модели MLP

model = mlp.MLP(mlp.sigmoid, mlp.dsigmoid, layers=[X.shape[1], 100, 100, len(np.unique(y))])
model.fit(X_train, y_train)

## Посмотрим точность

print(sum(np.argmax(model.predict(X_test), axis=1) == y_test) / y_test.shape[0])

## Выводы

Был написан и протестирован многослойный персептрон. Он показал неплохую точность, однако модели из специальных библиотек показывали более хорошую точность. Возможно увеличение количества итераций и изменение коэффициента обучения изменят результаты
