==> notebooks/8_vgg.ipynb <==
from tensorflow.python.client import device_lib
from tensorflow.python.keras.applications.inception_v3 import InceptionV3
from tensorflow.python.keras.applications.vgg16 import VGG16
from tensorflow.python.keras.models import Model
from tensorflow.python.keras.optimizer_v2.adam import Adam
from tensorflow.python.keras.preprocessing.image import ImageDataGenerator
from tensorflow.python.keras.layers.core import Dense

image_generator = ImageDataGenerator(rescale=1.0/255, rotation_range=15, zoom_range=0.15, width_shift_range=0.05, height_shift_range=0.05)

train_generator = image_generator.flow_from_directory("../cats_and_dogs/train", target_size=(224, 224),
                                                          class_mode= 'categorical', batch_size=1)

test_generator = image_generator.flow_from_directory("../cats_and_dogs/test", target_size=(224, 224),
                                                         class_mode= 'categorical', batch_size=1)

# Модель vgg16

vgg16 = VGG16(weights='imagenet', include_top=True)
vgg16.summary()

classification_layer = Dense(2, activation='softmax', name="images_prediction")
out = classification_layer(vgg16.layers[-2].output)
model = Model(vgg16.input, out)
model.summary()

for l, layer in enumerate(model.layers[:-1]):
    layer.trainable = False

model.layers[-1].trainable = True

model.compile(optimizer='adam', loss='categorical_crossentropy', metrics=['accuracy'])

model.fit_generator(train_generator, steps_per_epoch=2500, epochs=5, validation_data = test_generator,
          validation_steps = 200)

loss, accuracy = model.evaluate_generator(test_generator, steps=200)

print('Loss:', loss)
print('Accuracy:', accuracy)
